Create README.md
This file when run will enable the screen to turn on and off with attachment on Linux Mint Cinnamon 22. I run it as a login item.This file when run will enable the screen to turn on and off with attachment on Linux Mint Cinnamon 22. I run it as a login item.This file when run will enable the screen to turn on and off with attachment on Linux Mint Cinnamon 22. I run it as a login item.This file when run will enable the screen to turn on and off with attachment on Linux Mint Cinnamon 22. I run it as a login item.This file when run will enable the screen to turn on and off with attachment on Linux Mint Cinnamon 22. I run it as a login item.